Fix #151: Synchronize URLs with page labels - renamed /network to /now and /landingglobe to /2050
