+++ app/2050/popups/CapacityComparisionDrawer.tsx
