+++ app/2050/popups/ElectricityPriceComponent.tsx
