+++ app/2050/popups/GenerationmixDrawer.tsx
