+++ app/2050/popups/GlobeLegend.tsx
