+++ app/2050/popups/InvestementsPie.tsx
