+++ components/OverveiwComponents/hero.tsx
@@ -32,14 +32,14 @@ export const HeroSection = () => {
               className="w-full md:w-1/4 text-card-foreground dark:hover:bg-secondary hover:bg-ring"
               variant="outline"
             >
-              <Link href="/network">Now</Link>
+              <Link href="/now">Now</Link>
             </Button>
 
             <Button
               asChild
               className="w-full md:w-1/4 dark:bg-secondary bg-ring dark:text-card-foreground text-destructive-foreground hover:bg-popover-foreground dark:hover:bg-border "
             >
-              <Link href="/landingglobe">
+              <Link href="/2050">
                 2050
                 <ArrowRight className="size-5 ml-2 group-hover/arrow:translate-x-1 transition-transform hover:transform-45" />
               </Link>
